+++ src/BrokerSys/sys/Timeouts.h
@@ -24,6 +24,8 @@ namespace dcclite::broker::sys
 
 	auto constexpr NETWORK_DEVICE_PING_TIMEOUT = 2s;
 
+	auto constexpr NETWORK_DEVICE_FLOWRATE_INTERVAL = 10s;
+
 	auto constexpr TASK_DOWNLOAD_EEPROM_RETRY_TIMEOUT = 100ms;
 	auto constexpr TASK_DOWNLOAD_EEPROM_DOWNLOAD_WAIT = 25ms;
 	auto constexpr TASK_DOWNLOAD_EEPROM_WAIT_CONNECTION = 250ms;